// frontend/xacria-presenze/src/generated-client/model/calendarResponseDto.ts
/**
 * API Presenze-Pro
 * API used by the Presenze-Pro thesis project’s client
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CalendarEntry } from './calendarEntry';

/**
 * Oggetto in risposta a risposta di fetching di entires del calendario
 */
export interface CalendarResponseDto { 
    id?: string;
    entryType?: CalendarResponseDto.EntryTypeEnum;
    calendarEntry?: CalendarEntry;
}
export namespace CalendarResponseDto {
    export type EntryTypeEnum = 'WORKING_DAY' | 'REQUEST' | 'WORKING_TRIP' | 'AVAILABILITY';
    export const EntryTypeEnum = {
        WORKINGDAY: 'WORKING_DAY' as EntryTypeEnum,
        REQUEST: 'REQUEST' as EntryTypeEnum,
        WORKINGTRIP: 'WORKING_TRIP' as EntryTypeEnum,
        AVAILABILITY: 'AVAILABILITY' as EntryTypeEnum
    };
}